1933: Align DCLH feature with disbursements + add rejection scenarios
Refactor (DSTEW-1699 review):
- Rewrite duplicateChecksLegalHelp.feature in plain-English Gherkin
  matching the disbursements feature style (no HTTP/UI leakage).
- Consolidate step classes: fold BulkSubmissionApiSteps and
  LegacyDuplicateChecksCompatibilitySteps into a single
  LegalHelpDuplicateChecksSteps class.
- Fold BulkSubmissionLifecycleSupport PATCH helpers into
  BddApiStepSupport; remove the now-empty support class and its
  Spring bean wiring in BddBeansConfiguration.

Follow-up scenarios (close negative-path coverage gap):
- DCLH_12  Within-file duplicates rejected
- DCLH_13  Matching claim resubmitted inside the N-month duplicate
           window rejected (0 / 1 / 2 month examples)
- DCLH_14  Voiding the first claim unlocks resubmission 2 months later

No new step glue required; the shared vocabulary from
LegalHelpDisbursementsDuplicateChecksSteps is reused for all three
new scenarios via Cucumber's cross-class step registry.

Verified: ./gradlew :claims-data:service:bddTest
  SUCCESS: Executed 67 tests in 51.3s (0 failures)
